move `WarningModal` to `/common/modals`
