+++ clients/admin-ui/src/features/config-wizard/ScanResults.tsx
@@ -18,8 +18,8 @@ import {
 import { isErrorResult } from "~/features/common/helpers";
 import { useAPIHelper } from "~/features/common/hooks";
 import { useSystemOrDatamapRoute } from "~/features/common/hooks/useSystemOrDatamapRoute";
+import WarningModal from "~/features/common/modals/WarningModal";
 import { SystemsCheckboxTable } from "~/features/common/SystemsCheckboxTable";
-import WarningModal from "~/features/common/WarningModal";
 import {
   setSystemsToClassify,
   useUpsertSystemsMutation,
